:truck: setImageViewResource 위치 이동
